Refactoring for AQE
